requirements.txt
scipy >= 1
tensorflow_probability >= 0.9
